requirements.txt
openai
sentence-transformers
requests
feedparser
pandas
yfinance
python-dotenv
